@@ -9,7 +9,7 @@
   "homepage": "http://www.phpactiverecord.org/",
   "license": "MIT",
   "require": {
-    "php": ">=7.1.0"
+    "php": ">=8.1"
   },
   "require-dev": {
     "phpunit/phpunit": "4.*",
