@@ -1,11 +1,10 @@
 <?php
-if (!defined('PHP_VERSION_ID') || PHP_VERSION_ID < 70200)
-    die('PHP ActiveRecord requires PHP 7.2 or higher');
 
-define('PHP_ACTIVERECORD_VERSION_ID', '1.2.1');
+if (!defined('PHP_VERSION_ID') || PHP_VERSION_ID < 80100) {
+	die('PHP ActiveRecord requires PHP 8.1 or higher');
+}
 
-if (!defined('PHP_ACTIVERECORD_AUTOLOAD_PREPEND'))
-    define('PHP_ACTIVERECORD_AUTOLOAD_PREPEND', true);
+const PHP_ACTIVERECORD_VERSION_ID = '2.0.1';
 
 require __DIR__ . '/lib/Singleton.php';
 require __DIR__ . '/lib/Config.php';
@@ -23,27 +22,3 @@ require __DIR__ . '/lib/Inflector.php';
 require __DIR__ . '/lib/CallBack.php';
 require __DIR__ . '/lib/Exceptions.php';
 require __DIR__ . '/lib/Cache.php';
-
-if (!defined('PHP_ACTIVERECORD_AUTOLOAD_DISABLE'))
-	spl_autoload_register('activerecord_autoload');
-
-function activerecord_autoload($class_name)
-{
-    $path = ActiveRecord\Config::instance()->get_model_directory();
-    $root = realpath(isset($path) ? $path : '.');
-
-    if (($namespaces = ActiveRecord\get_namespaces($class_name))) {
-        $class_name = array_pop($namespaces);
-        $directories = array();
-
-        foreach ($namespaces as $directory)
-            $directories[] = $directory;
-
-        $root .= DIRECTORY_SEPARATOR . implode(DIRECTORY_SEPARATOR, $directories);
-    }
-
-    $file = "$root/$class_name.php";
-
-    if (file_exists($file))
-        require_once $file;
-}
